api/pyproject.toml
[tool.poetry]
name = "business-ar-api"
version = "1.1.0"
description = ""
authors = ["thorwolpert <thor@wolpert.ca>"]
license = "BSD 3-Clause"
readme = "README.md"
packages = [{include = "business_ar_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12.2"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.4"
launchdarkly-server-sdk = "^8.3.0"
python-dotenv = "^1.0.1"
sentry-sdk = "^1.44.1"
flask-babel = "^3.1.0"
psycopg2 = "^2.9.5"
flask-jwt-oidc = "^0.7.0"
gunicorn = "^21.2.0"
flask-cors = "^4.0.0"
jsonschema = {extras = ["format"], version = "^4.20.0"}
pg8000 = "^1.31.2"
Jinja2 = "^3.1.2"

[tool.poetry.group.test.dependencies]
black = "^24.4.2"
freezegun = "^1.4.0"
hypothesis = "^6.100.0"
pexpect = "^4.9.0"
pytest = "^7.4.4"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
requests-mock = "^1.12.1"
nanoid = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
faker = "^24.4.0"
lovely-pytest-docker = "^0.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
